requirements.txt
Flask==2.3.3
supabase==1.0.3
python-dotenv==1.0.0
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.6